Completed day 2

+++ src/input.rs
@@ -2,7 +2,7 @@ use std::fs::File;
 use std::io::prelude::*;
 use std::io::BufReader;
 
-pub fn get_day_input(day: i32) -> Vec<String> {
+pub fn get_day_input(day: usize) -> Vec<String> {
   let file_name = String::from("resources/day_") + &day.to_string() + ".txt";
 
   let file = File::open(file_name).expect("Day input doesn't exist");
